renamed the flask app again, to ui_server.py
